+++ pmc_automation_tools/common/utils.py
@@ -1,18 +1,26 @@
 from datetime import datetime, date, timedelta, timezone
+import pytz
 import os
 import sys
 import json
 import csv
 from warnings import warn
 import logging
-from typing import Union
+from logging.handlers import MemoryHandler
+
+from typing import Union, Generator, List, Dict, Any
 from openpyxl import load_workbook
 
 DEFAULT_FORMATTER = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+STDOUT_FORMATTER = "[%(asctime)s][%(filename)s:%(lineno)s][%(funcName)20s()] %(message)s"
+LOG_FORMATS = {
+    "DAILY": "%Y_%m_%d_",
+    "MONTHLY": "%Y_%m_"
+    }
+
 def debug_logger(level=logging.NOTSET):
     logger = logging.getLogger(__name__)
-    FORMAT = "[%(asctime)s][%(filename)s:%(lineno)s][%(funcName)20s()] %(message)s"
-    logging.basicConfig(format=FORMAT)
+    logging.basicConfig(format=STDOUT_FORMATTER)
     logger.setLevel(level)
     return logger
 
@@ -30,6 +38,8 @@ def frozen_check():
     # Running in a normal Python environment
         bundle_dir = os.path.dirname(os.path.abspath(__file__))
     return bundle_dir
+
+
 def debug_dump_variables(obj):
     """
     Dumps variables of provided object to a log file.
@@ -45,10 +55,12 @@ def debug_dump_variables(obj):
         obj.debug_logger.debug(k, v)
         obj.dump_logger.debug(f'{k} : {v}')
 
+
 def get_case_insensitive_key_value(input_dict, key):
     return next((value for dict_key, value in input_dict.items() if dict_key.lower() == key.lower()), None)
 
-def create_batch_folder(root='', batch_code=None, include_time=False, test=False, batch_prefix=None) -> str:
+
+def create_batch_folder(root:str='', batch_code:str=None, include_time:bool=False, test:bool=False, batch_prefix:bool=None) -> str:
     """
     Used to set up a batch folder to store any log files or screenshots during an automation run.
     
@@ -77,44 +89,57 @@ def create_batch_folder(root='', batch_code=None, include_time=False, test=False
     batch_folder = os.path.join(root, 'batch_codes', db, folder_name)
     os.makedirs(batch_folder, exist_ok=True)
     return batch_folder
-def setup_logger(name, log_file='log.log', file_format='DAILY',
-                     level=logging.DEBUG, formatter=DEFAULT_FORMATTER, root_dir=None) -> logging.Logger:
+
+
+def setup_logger(name:str,
+                 log_file:str='log.log',
+                 file_format:str='DAILY',
+                 level:Union[int, str]=logging.DEBUG,
+                 formatter:str=DEFAULT_FORMATTER,
+                 root_dir:str=None,
+                 flush_level:Union[int, str]=logging.ERROR,
+                 write_stdout:bool=True) -> logging.Logger:
     """
-    Setup a logging file.
+    Setup a logger with a memory buffer that flushes on errors, and optionally outputs to stdout.
 
     Parameters:
-
     - name: logger name
     - log_file: filename for the log file.
     - file_format: "DAILY" | "MONTHLY" | "". Will be combined with the log_file filename provided.
     - level: log level for the logger. logging module levels.
-    - formatter: logging formatter
-    - root_dir: root directory to store the log file
+    - formatter: logging formatter.
+    - root_dir: root directory to store the log file.
+    - flush_level: log level to trigger flushing to the file.
+    - write_stdout: if True, logs will also be printed to stdout.
 
     Default formatter: %(asctime)s - %(name)s - %(levelname)s - %(message)s
     """
-    _name = str(name)
-    _file_format = str(file_format).upper()
-    today = datetime.now()
-    _formatter = logging.Formatter(formatter)
-    if _file_format == 'DAILY':
-        log_date = today.strftime("%Y_%m_%d_")
-    elif _file_format == 'MONTHLY':
-        log_date = today.strftime("%Y_%m_")
-    else:
-        log_date = ''
-    _log_file = log_date + log_file
-    if root_dir:
-        _log_file = os.path.join(root_dir, _log_file)
-    handler = logging.FileHandler(_log_file, mode='a', encoding='utf-8')
-    handler.setFormatter(_formatter)
-    logger = logging.getLogger(_name)
+    date_format = LOG_FORMATS.get(file_format.upper(), "")
+    log_date = datetime.now().strftime(date_format) if date_format else ""
+    log_path = os.path.join(root_dir, log_date + log_file) if root_dir else log_date + log_file
+
+    file_handler = logging.FileHandler(log_path, mode='a', encoding='utf-8')
+    file_handler.setFormatter(logging.Formatter(formatter))
+
+    memory_handler = MemoryHandler(
+        capacity=1024,
+        flushLevel=flush_level,
+        target=file_handler
+    )
+    
+    logger = logging.getLogger(name)
     logger.setLevel(level)
+    
     if not logger.hasHandlers():
-        logger.addHandler(handler)
+        logger.addHandler(memory_handler)
+        if write_stdout:
+            stream_handler = logging.StreamHandler()
+            stream_handler.setFormatter(logging.Formatter(STDOUT_FORMATTER))
+            stream_handler.setLevel(level)
+            logger.addHandler(stream_handler)
     return logger
 
-def read_updated(in_file, obj_type=None) -> Union[list, dict]: 
+def read_updated(in_file:str, obj_type:Union[dict, list]=None) -> Union[List[Dict[str, str]],Any]: 
     
     """
     Read in a file of already updated records.
@@ -147,7 +172,7 @@ def read_updated(in_file, obj_type=None) -> Union[list, dict]:
     return updated_records
 
 
-def _read_excel(file_path):
+def _read_excel(file_path:str) -> List[Dict[str, str]]:
     """
     Reads the contents of the first sheet in an Excel (.xlsx) file.
 
@@ -182,7 +207,7 @@ def _read_excel(file_path):
     return data
 
 
-def save_updated_overwrite(in_file, obj):
+def save_updated_overwrite(in_file:str, obj:dict) -> None:
     """
     Save a file containing a list of already processed records.
 
@@ -205,21 +230,20 @@ def save_updated_overwrite(in_file, obj):
             raise TypeError('File name provided is not an expected type of json or csv.')
         
 
-def save_updated(in_file, obj):
+def save_updated(in_file:str, obj:Union[dict, list]) -> None:
     """
     Append to a file containing a list of already processed records.
 
     Parameters:
     
     - in_file: file to use to save
-    - obj: json object to write to file. Expects a list containing dictionaries.
+    - obj: json object to write to file.
     """
     if not obj:
         return
-    if not isinstance(obj, list):
-        obj = [obj]
-    if len(obj) > 1:
+    if isinstance(obj, list):
         return save_updated_overwrite(in_file, obj)
+    obj = [obj]
     _file_type = in_file.split('.')[-1].lower()
     
     # Append JSON data
@@ -253,7 +277,7 @@ def save_updated(in_file, obj):
         raise TypeError('File name provided is not an expected type of json or csv.')
     
 
-def plex_date_formatter(*args: datetime|int, date_offset=0, tz_convert=True):
+def plex_date_formatter(*args: datetime|int, date_offset:int=0, tz_convert:bool=True) -> str:
     """
     Takes 'normal' date formats and converts them to a Plex web service format (ISO format)
     Can also take a single datetime object.
@@ -269,12 +293,13 @@ def plex_date_formatter(*args: datetime|int, date_offset=0, tz_convert=True):
     else:
         _date = datetime(*args)
     if tz_convert:
-        _date = _date.astimezone(datetime.now(timezone.utc).tzinfo)
+        ny_tz = pytz.timezone("America/New_York")
+        _date = ny_tz.localize(_date).astimezone(timezone.utc)
     _date += timedelta(days=date_offset)
     f_date = _date.strftime('%Y-%m-%dT%H:%M:%SZ')
     return f_date
 
 
-def chunk_list(lst, chunk_size):
+def chunk_list(lst:list, chunk_size:int) -> Generator[list, None, None]:
     for i in range(0, len(lst), chunk_size):
         yield lst[i:i + chunk_size]