+++ CHANGELOG.md
@@ -24,6 +24,10 @@ Added `plex_date_formatter()` to `__init__.py`.
 
 `common.utils.save_updated()` Added input to overwrite the file contents regardless of the input object.
 
+Added support for xlsm files in `read_updated()`.
+
+Added support for specifying sheet name when reading Excel files in `read_updated()`.
+
 ## Fixed
 
 Fixed `api.ux.datasource.type_reconcile()` when dealing with datetime values.
@@ -45,6 +49,8 @@ Changed `common.utils.save_updated()` to append single dictionary values to the 
 Rewrote `common.utils.setup_logger()` to be less complicated and include memoryhandler/streamhandler support.
 
 Added support to return the full item in `DataSourceResponse.get_response_attribute()` using 'ALL' as the input variable.
+
+Changed `DataSourceResponse.get_response_attribute()` to return `None` if there are no matches.
 
 # 0.4.0 [2024-11-1]